+++ sketch-core/src/commonMain/kotlin/com/github/panpf/sketch/resize/PrecisionDecider.kt
@@ -59,26 +59,22 @@ data class FixedPrecisionDecider(private val precision: Precision) : PrecisionDe
  *
  * Note: The precision parameter can only be [Precision.EXACTLY] or [Precision.SAME_ASPECT_RATIO].
  */
-class LongImageClipPrecisionDecider constructor(
+class LongImagePrecisionDecider constructor(
     val longImage: Precision = Precision.SAME_ASPECT_RATIO,
     val otherImage: Precision = Precision.LESS_PIXELS,
     val longImageDecider: LongImageDecider = LongImageDecider(),
 ) : PrecisionDecider {
 
-    override val key: String by lazy { "LongImageClip($longImage,$otherImage,${longImageDecider.key})" }
+    override val key: String by lazy { "LongImage($longImage,$otherImage,${longImageDecider.key})" }
 
     override fun get(imageSize: Size, targetSize: Size): Precision {
         val isLongImage = longImageDecider.isLongImage(imageSize, targetSize)
         return if (isLongImage) longImage else otherImage
     }
 
-    override fun toString(): String {
-        return "LongImageClipPrecisionDecider(precision=$longImage, otherPrecision=$otherImage, longImageDecider=$longImageDecider)"
-    }
-
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
-        if (other !is LongImageClipPrecisionDecider) return false
+        if (other !is LongImagePrecisionDecider) return false
         if (longImage != other.longImage) return false
         if (otherImage != other.otherImage) return false
         if (longImageDecider != other.longImageDecider) return false
@@ -90,5 +86,9 @@ class LongImageClipPrecisionDecider constructor(
         result = 31 * result + otherImage.hashCode()
         result = 31 * result + longImageDecider.hashCode()
         return result
+    }
+
+    override fun toString(): String {
+        return "LongImagePrecisionDecider(longImage=$longImage, otherImage=$otherImage, longImageDecider=$longImageDecider)"
     }
 }